Rename module to allow non-IOC features to be included
There is some overlap between managing tools that are shipping as
part of EPICS Base like caRepeater and managing IOCs. In some cases
the same packages might need to be installed. Folding this into a
single module allows for code to be reused easily and prevents
potential compatibility issues between the different Puppet modules.